correction de certains trucs

+++ VectorDrawPRO/VectorDrawPRO/Code/Models/Shape.cs
@@ -4,10 +4,10 @@ namespace VectorDrawPRO.Code.Models;
 
 public abstract class Shape
 {
-    public int X { get; set; }
-    public int Y { get; set; }
-    public int Width { get; set; }
-    public int Height { get; set; }
+    public double X { get; set; }
+    public double Y { get; set; }
+    public double Width { get; set; }
+    public double Height { get; set; }
 
     public abstract void Draw(Canvas canvas);
 }
